bfilter/src/requirements.txt
python-dotenv
numpy
pandas
scikit-learn
Flask
gunicorn
requests
google-auth
joblib
google-cloud-pubsub
